pyproject.toml
[project]
name = "swara"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lingua-language-detector>=2.1.1",
    "pandas>=2.3.2",
    "phonemizer>=3.3.0",
    "accelerate>=1.10.1",
    "einops>=0.8.1",
    "einops-exts>=0.0.4",
    "librosa>=0.11.0",
    "matplotlib>=3.10.6",
    "monotonic-align",
    "munch>=4.0.0",
    "nltk>=3.9.1",
    "pydub>=0.25.1",
    "pyyaml>=6.0.2",
    "soundfile>=0.13.1",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
    "typing>=3.10.0.0",
    "typing-extensions>=4.15.0",
    "wandb>=0.21.3",
    "singleton-decorator>=1.0.0",
    "datasets>=4.0.0",
    "sacremoses>=0.1.1",
    "pebble>=5.1.3",
    "ipykernel>=6.30.1",
    "tensorboard>=2.20.0",
    "openpyxl>=3.1.5",
    "editdistance>=0.8.1",
]

[project.scripts]
phonemize-text = "phonemize.__main__:main"
phonemize-preprocess = "phonemize.preprocess.cli:main"

[tool.uv.sources]
monotonic-align = { git = "https://github.com/resemble-ai/monotonic_align.git" }
